+++ R/makeMetageneDataFrame.R
@@ -25,7 +25,9 @@
 #' # Make a list of samples
 #' sampleNames <- head(LETTERS, 2)
 #' binCenters <- seq_len(nWindows) - nWindows/2
-#' seList <- lapply(sampleNames, function(x){makeExperimentFromDeeptools(tf, col.names=binCenters)})
+#' seList <- lapply(sampleNames, function(x){
+#'     makeExperimentFromDeeptools(tf, col.names=binCenters)
+#' })
 #' names(seList) <- sampleNames
 #'
 #' # Split each sample into gene subsets