+++ tests/testthat/setup-library.R
@@ -1,0 +1,3 @@
+library(rtracklayer)
+library(GenomicRanges)
+library(SummarizedExperiment)